requirements.txt
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
channels==4.0.0
channels-redis==4.0.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==41.0.4
cupshelpers==1.0
daphne==4.0.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.5
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.11.0
django-crispy-forms==2.0
django-import-export==3.3.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-taggit==4.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
future==0.18.2
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
pandas==2.1.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.0.1
protobuf==3.12.4
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2021.3
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
systemd-python==234
tablib==3.5.0
toml==0.10.2
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
zope.interface==6.0
